（# 71）code 添加其他人的个人主页进入关注粉丝列表的逻辑

+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/follow_fans/FollowFansListContract.java
@@ -48,9 +48,9 @@ public interface FollowFansListContract {
          * @param userId     用户id
          * @param pageType   详见FollowFansListFragment.class定义的页面类型
          */
-        void requestNetData(Long maxId, boolean isLoadMore, int userId, int pageType);
+        void requestNetData(Long maxId, boolean isLoadMore, long userId, int pageType);
 
-        List<FollowFansBean> requestCacheData(Long maxId, boolean isLoadMore, int userId, int pageType);
+        List<FollowFansBean> requestCacheData(Long maxId, boolean isLoadMore, long userId, int pageType);
 
         /**
          * 关注用户
